+++ demo/app/main-page.ts
@@ -5,8 +5,15 @@ export function onLoaded (args) {
 
     page.bindingContext = fromObject({
         examples: [
-            { title: 'NS Chart', moduleName: 'examples/NSChart' },
-            { title: 'NS UI Chart', moduleName: 'examples/NSUIChart' },
+            { title: 'Add data points in realtime', moduleName: 'examples/RealTime' },
+            { title: 'Simple Bar Chart', moduleName: 'examples/BarChart' },
+            { title: 'Simple Bubble Chart', moduleName: 'examples/BubbleChart' },
+            { title: 'Simple CandleStick Chart', moduleName: 'examples/CandleStickChart' },
+            { title: 'Simple Horizontal Bar Chart', moduleName: 'examples/HorizontalBarChart' },
+            { title: 'Simple Line Chart', moduleName: 'examples/LineChart' },
+            { title: 'Simple Pie Chart', moduleName: 'examples/PieChart' },
+            { title: 'Simple Radar Chart', moduleName: 'examples/RadarChart' },
+            { title: 'Simple Scatter Chart', moduleName: 'examples/ScatterChart' },
         ],
     });
 };
@@ -15,5 +22,10 @@ export function goToExample  (args) {
     const page = args.object.page;
 
     console.log('goToExample');
-    page.frame.navigate(page.bindingContext.examples[args.index].moduleName);
+    const entry = page.bindingContext.examples[args.index];
+    page.frame.navigate({
+        moduleName: entry.moduleName,
+        context: entry,
+        animated: true
+    });
 };