+++ plugin/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/nativescript-community/ui-chart/compare/v1.0.0...v1.0.1) (2020-09-30)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+
+
+
+
+
 # [1.0.0](https://github.com/Akylas/nativescript-chart/compare/v0.0.37...v1.0.0) (2020-09-29)
 
 **Note:** Version bump only for package @nativescript-community/ui-chart